histogram added

+++ back-end/server.py
@@ -7,21 +7,22 @@ from fastapi.responses import JSONResponse, FileResponse
 from fastapi.staticfiles import StaticFiles
 import pandas as pd 
 
-from new_utils import renew_model, renew_make_gradcam
+from new_utils import renew_model, renew_make_gradcam, visual_histogram
 import shutil, zipfile, uuid
 
 class textField(BaseModel) :
   text: str
 
-user_images_folder = f"images" # = f"user_images"
-model_folder = f"model" # tf_keras_vgg16_mura_model.h5
+user_images_folder = f"user_images"
+model_folder = f"model"
 save_heatmap = f"heatmap"
 
 # Arbitrary location for csv. File name involved.
 csv_location = f"test.csv"
 
 # Arbitrary White (Blank) image
-white_image_loc = f"white.png"
+white_image_loc = f"heatmap/white.png"
+histogram_save_location = f"heatmap/histogram"
 
 app = FastAPI()
 
@@ -34,6 +35,7 @@ app.add_middleware(
 )
 
 app.mount("/heatmap", StaticFiles(directory="heatmap"), name="heatmap")
+# app.mount("/white", StaticFiles(directory="white"), name="white")
 
 @app.post("/textreturn")
 def text_return(data: dict):
@@ -51,6 +53,8 @@ async def upload_images(files: List[UploadFile]):
     if os.path.exists(user_images_folder):
         for filename in os.listdir(user_images_folder):
             file_path = os.path.join(user_images_folder, filename)
+            if os.path.isdir(file_path):
+                continue
             if os.path.isfile(file_path):
                 os.unlink(file_path)
     else:
@@ -68,14 +72,8 @@ async def upload_images(files: List[UploadFile]):
 '''
 Find the result images and return 4 image urls.
 '''
-def select_images(csv_location, white_image_loc,  class_id : int = 0 , column_id : int =1) -> list:
+def select_images(csv_location, white_image_loc,  class_id : int = 1 , column_id : int =1) -> list:
     dataframe = pd.read_csv(csv_location)
-    max_column_id = max(dataframe.loc[dataframe['prediction'] == class_id, 'column_id'].tolist())
-    
-    # 혹 coulmn_id 가 실제 column의 개수보다 많이 들어오면, 메세지를 출력한 다음, 마지막으로 값으로 수정함. 
-    if column_id > max_column_id : 
-        print("column id is bigger than max column_id")
-        column_id = max_column_id
     selected_rows= dataframe[(dataframe['prediction'] == class_id) & (dataframe['column_id'] == column_id)]
 
     heatmap_address = selected_rows['heatmap_path'].tolist() 
@@ -83,6 +81,17 @@ def select_images(csv_location, white_image_loc,  class_id : int = 0 , column_id
         heatmap_address.append(white_image_loc)
     
     return heatmap_address
+
+def load_histogram(class_id, histogram_save_location, save_heatmap) : 
+    histogram_path = os.path.join(histogram_save_location, f'histogram_{class_id}.png').replace('\\', '/')
+
+    # Check if image exists.
+    if os.path.exists(histogram_path):
+        # Return the image path.
+        return histogram_path
+    else:
+        return os.path.join(save_heatmap, 'No_Data.png').replace('\\', '/')
+
 
 @app.post("/upload-model")
 async def upload_model(zipFile: UploadFile):
@@ -140,21 +149,34 @@ async def run_gradcam():
     # 1) renew save_heatmap folder 
     # 2) make gradcam & save heatmap in save_heatmap folder
     # 3) save csv with infomation at 'csv_location' 
-    renew_make_gradcam(model_location, user_images_folder, save_heatmap, csv_location)
+    num_class = renew_make_gradcam(model_location, user_images_folder, save_heatmap, csv_location)
+    visual_histogram(num_class, csv_location, save_folder = histogram_save_location)
+    
+    df = pd.read_csv(csv_location)
 
+    max_column_id = [] 
+    for c_id in range(num_class) : 
+        c_column_id = df.loc[df['prediction'] == c_id + 1, 'column_id'].tolist()
+        if len(c_column_id) == 0 : 
+            class_max_column_id = 0 
+        else : 
+            class_max_column_id = max(c_column_id)
+        
+        max_column_id.append(class_max_column_id)
+        
     # Initialize global column/ class id.
     global global_column_id 
     global_column_id = 1
     
     global global_class_id
-    global_class_id = 0 
+    global_class_id = 1
 
     result = select_images(csv_location, white_image_loc)
-    default_maximum_value = [3, 3, 3]
+    histogram_path = load_histogram(class_id = 1, histogram_save_location= histogram_save_location, save_heatmap= save_heatmap)
     
     result = [img[8:] for img in result]
-    print(result)
-    return {'image_paths': result, 'max_value': default_maximum_value}
+    print(result, max_column_id)
+    return {'image_paths': result, 'max_value': max_column_id, 'histogram': histogram_path[8:]}
 
 @app.post("/next-button")
 async def next_button():
@@ -170,11 +192,11 @@ async def prev_button():
 
 @app.post("/class-dropdown")
 async def prev_button(class_num: int):
-    global global_class_num
-    global_class_num = class_num
+    global_class_id = class_num
+    global_column_id = 0
     
     result = select_images(csv_location, white_image_loc, global_class_id, global_column_id)
     return {'image_paths': result}
 
 if __name__=='__main__':
-    uvicorn.run(app, host='0.0.0.0', port = 8000)
+    uvicorn.run(app, host='110.76.86.172', port = 8000)
